add bank repository, service, controller and dtos

+++ src/main/java/com/taulantavdiu/banksystem/dtos/AccountResponseDto.java
@@ -2,6 +2,7 @@ package com.taulantavdiu.banksystem.dtos;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Builder
@@ -9,7 +10,7 @@ public record AccountResponseDto(
         UUID id,
         String firstName,
         String lastName,
-        Double balance
+        BigDecimal balance
 
 ) {
 }
